Remove 80 characters per line recommendation.
This commit would necessitate a bump to the major version as it reverts
a previous recommendation.

An 80 character limit should not be imposed on the code, since Puppet
lacks a method for line separation, such as \ in bash. It is highly
regular to have lines that exceed 80 characters, especially dealing with
exec and cron resources, validate_re(), and returning useful messages
with fail().